Updated healthbar visuals
Updated the health and shield bar numbers to be more visible and make both face the camera on the same plane
