+++ Vivarium/Assets/Scripts/UI/FaceCamera.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FaceCamera : MonoBehaviour
+{
+    private Camera _camera;
+    private void Start()
+    {
+        _camera = Camera.main;
+    }
+
+    private void Update()
+    {
+        transform.rotation = _camera.transform.rotation;
+    }
+}
