+++ Vivarium/Assets/Scripts/UI/FaceCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e6d3240cde2fc8445885084bdfef0ad3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
