+++ Vivarium/Assets/Scripts/UI/HealthBar.cs
@@ -11,17 +11,6 @@ public class HealthBar : MonoBehaviour
 
     private float _maxHealth = 0f;
     private float _currentHealth = 0f;
-    private Camera _camera;
-
-    private void Start()
-    {
-        _camera = Camera.main;
-    }
-
-    private void Update()
-    {
-        transform.rotation = _camera.transform.rotation;
-    }
 
     public void SetMaxHealth(float maxHealth)
     {
